Separated out executor values for each data structure
