+++ src/main/kotlin/com/manimdsl/runtime/VirtualMachine.kt
@@ -5,11 +5,17 @@ import com.manimdsl.ExitStatus
 import com.manimdsl.errorhandling.ErrorHandler.addRuntimeError
 import com.manimdsl.frontend.*
 import com.manimdsl.linearrepresentation.*
+import com.manimdsl.runtime.datastructures.array.Array2DValue
 import com.manimdsl.runtime.datastructures.array.ArrayExecutor
+import com.manimdsl.runtime.datastructures.array.ArrayValue
 import com.manimdsl.runtime.datastructures.binarytree.BinaryTreeExecutor
+import com.manimdsl.runtime.datastructures.binarytree.BinaryTreeNodeValue
+import com.manimdsl.runtime.datastructures.binarytree.BinaryTreeValue
+import com.manimdsl.runtime.datastructures.binarytree.NullValue
+import com.manimdsl.runtime.datastructures.makeConstructorNode
 import com.manimdsl.runtime.datastructures.stack.StackExecutor
+import com.manimdsl.runtime.datastructures.stack.StackValue
 import com.manimdsl.runtime.utility.getBoundaries
-import com.manimdsl.runtime.utility.makeConstructorNode
 import com.manimdsl.runtime.utility.wrapCode
 import com.manimdsl.runtime.utility.wrapString
 import com.manimdsl.shapes.SubtitleBlockShape
